% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpcde_fns.R
\name{b_x}
\alias{b_x}
\title{bx = constants for each data and evaluation point pair}
\usage{
b_x(datavec, s_mat, e_vec, q, kernel_type)
}
\arguments{
\item{s_mat}{s_hat matrix.}

\item{q}{polynomial order.}

\item{kernel_type}{kernel function choice.}

\item{data_vec}{dataset, vector.}
}
\value{
vector of products for each data point.
}
\description{
Function for estimating the constants in the estimation formula
}
\keyword{internal}
