% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{basis_vec}
\alias{basis_vec}
\title{Unit basis vector}
\usage{
basis_vec(x, p, mu)
}
\arguments{
\item{x}{sample input scalar or vector.}

\item{p}{polynomial order.}

\item{mu}{derivative order.}
}
\value{
Vector of appropriate length with ones corresponding to entries of order \code{mu}.
}
\description{
Function to generate unit basis vector according to polynomial order
and derivative order. This function returns unit vector that is the same size
as the vector returned by \code{poly_base(x, p)}.
}
\examples{
basis_vec(x = 2, p = 5, mu = 1)
}
