% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{basis_vecC}
\alias{basis_vecC}
\title{unit basis vector in C++}
\usage{
basis_vecC(x, p, mu)
}
\arguments{
\item{x}{a number or vector.}

\item{p}{a number.}

\item{mu}{a number.}
}
\value{
unit vector of appropriate length with ones corresponding to entries of order \code{mu}.
}
\description{
unit vector basis function.
}
\keyword{internal}
