% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpbwcde_fns.R
\name{bw_rot}
\alias{bw_rot}
\title{ROT Bandwidth selection}
\usage{
bw_rot(y_data, x_data, y_grid, x, p, q, mu, nu, kernel_type)
}
\arguments{
\item{y_data}{Numeric matrix/data frame, the raw data of independent.}

\item{x_data}{Numeric matrix/data frame, the raw data of covariates.}

\item{y_grid}{Numeric vector, the evaluation points.}

\item{x}{Numeric, specifies the evaluation point(s) in the x-direction.}

\item{p}{Integer, polynomial order.}

\item{q}{Integer, polynomial order.}

\item{mu}{Integer, order of derivative.}

\item{nu}{Integer, order of derivative.}

\item{kernel_type}{String, the kernel.}
}
\value{
bandwidth sequence
}
\description{
Internal Function
}
\keyword{internal}
