% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpbwcde_fns.R
\name{c_exact}
\alias{c_exact}
\title{C matrix
Internal Function}
\usage{
c_exact(lower, upper, eval_pt, m, p, kernel_type)
}
\arguments{
\item{eval_pt}{evaluation point.}

\item{m}{derivative order.}

\item{p}{Nonnegative integer, polynomial order.}

\item{kernel_type}{String, kernel type}
}
\value{
a (p+1)-by-1 matrix
}
\description{
C matrix
Internal Function
}
\keyword{internal}
