% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{c_x}
\alias{c_x}
\title{c_x vector (Internal Function)}
\usage{
c_x(x_data, eval_pt, m, q, h, kernel_type)
}
\arguments{
\item{x_data}{data of covariates.}

\item{eval_pt}{evaluation point.}

\item{m}{order of polynomial.}

\item{q}{maximum degree for x.}

\item{h}{bandwidth.}

\item{kernel_type}{type of kernel function.}
}
\description{
c_x vector generated as described in main paper.
}
\keyword{internal}
