% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{c_y}
\alias{c_y}
\title{c_y vector (Internal Function)}
\usage{
c_y(y_data, eval_pt, m, p, h, kernel_type)
}
\arguments{
\item{y_data}{set of data points.}

\item{eval_pt}{evaluation point.}

\item{m}{order of derivative.}

\item{p}{maximum degree for y.}

\item{h}{bandwidth.}

\item{kernel_type}{type of kernel function.}
}
\value{
vector object, c_y.
}
\description{
c_x vector generated as described in main paper.
}
\keyword{internal}
