% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpbwcde_methods.R
\name{coef.lpbwcde}
\alias{coef.lpbwcde}
\title{Coef method for local polynomial density bandwidth selection}
\usage{
\method{coef}{lpbwcde}(object, ...)
}
\arguments{
\item{object}{Class "lpbwcde" object, obtained by calling \code{\link{lpbwcde}}.}

\item{...}{Other arguments.}
}
\value{
\item{Matrix}{A matrix containing y_grid points and selected bandwidths.}
}
\description{
The coef method for local polynomial density bandwidth selection objects.
}
\examples{
set.seed(42)
n <- 100
x_data <- as.matrix(rnorm(n, mean = 0, sd = 1))
y_data <- as.matrix(rnorm(n, mean = 0, sd = 1))
# bandwidth selection
y_grid <- stats::quantile(y_data, seq(from = 0.1, to = 0.9, by = 0.1))
model2 <- lpcde::lpbwcde(y_data = y_data, x_data = x_data, x = 0,
                         y_grid = y_grid, bw_type = "mse-rot")
coef(model2)

}
\seealso{
\code{\link{lpbwcde}} for data-driven bandwidth selection.

Supported methods: \code{\link{coef.lpbwcde}}, \code{\link{print.lpbwcde}}, \code{\link{summary.lpbwcde}}.
}
\author{
Matias D. Cattaneo (maintainer). \email{matias.d.cattaneo@gmail.com}.

Rajita Chandak. \email{rajita.chandak@gmail.com}.

Michael Jansson. \email{michael.jansson.berkeley@gmail.com}.

Xinwei Ma. \email{xinweima.pku@gmail.com}.
}
