% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{dgp}
\alias{dgp}
\title{Data generating function}
\usage{
dgp(n, xmin, xmax, dgp_type)
}
\arguments{
\item{n}{number of datapoints.}

\item{xmin}{minimum value of covariate.}

\item{xmax}{maximum value of covariate.}

\item{dgp_type}{joint distribution from which samples are generated.}
}
\value{
dataset of size n.
}
\description{
Function that generates dataset of prescribed size and distribution.
Covariates are univariate.\cr
Types of dgp:\cr
X: uniform(xmin, xmax), Y: normal(x,1),\cr
X: uniform(xmin, xmax), Y: uniform(min(X), max(X)),\cr
X: normal(0,1), Y: normal(x,1),\cr
X: uniform(\{0.5, 1.5\}), Y: normal(0.5, 1)+normal(1.5, 1),
}
\keyword{internal}
