% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{dgp_eval}
\alias{dgp_eval}
\title{PDF evaluation}
\usage{
dgp_eval(x, x_star, dgp_type)
}
\arguments{
\item{x}{conditioned covariate value.}

\item{x_star}{evaluation point.}

\item{dgp_type}{true joint distribution.}
}
\value{
value between 0 and 1.
}
\description{
Function that evaluates true pdf at conditioned value based on joint distribution.
}
\keyword{internal}
