% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{int_val}
\alias{int_val}
\title{Lp integral (Internal Function)}
\usage{
int_val(l, a, b, kernel_type)
}
\arguments{
\item{l}{Degree of polynomial being integrated.}

\item{a}{Lower limit of integration.}

\item{b}{Upper limit of integration.}

\item{kernel_type}{Type of kernel function. Choose from "uniform", "triangular", "epanechnikov".}
}
\value{
Value of integral.
}
\description{
Local polynomial integral evaluation
calculation of elements of S_y and middle integral (evaluating integral at end points).
}
\keyword{internal}
