% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{int_valC}
\alias{int_valC}
\title{lp integral evaluation in C++ (Internal Function)}
\usage{
int_valC(l, a, b, kernel_type)
}
\arguments{
\item{l}{degree of polynomial being integrated.}

\item{a}{lower limit of integration.}

\item{b}{upper limit of integration.}

\item{kernel_type}{type of kernel function. Choose from "uniform", "triangular", "epanechnikov".}
}
\value{
value of integral.
}
\description{
local polynomial integral evaluation calculation of elements
of S_y and middle integral (evaluating integral at end points)
}
\keyword{internal}
