% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kernel_evalC}
\alias{kernel_evalC}
\title{Kernel function C++ (Internal Function)}
\usage{
kernel_evalC(x, kernel_type)
}
\arguments{
\item{x}{evaluation point.}

\item{kernel_type}{kernel function type.}
}
\value{
kernel evaluated at \code{x}.
}
\description{
Multivariate product kernel evaluation function. Compatible with uniform,
triangular and epanechnikov kernels.
}
\keyword{internal}
