% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mvec}
\alias{mvec}
\title{Polynomial order vector}
\usage{
mvec(n, d)
}
\arguments{
\item{n}{Total value of each combination}

\item{d}{Maximum length of combinations}
}
\description{
Generates list of all combinations
of length less than or equal to d of numbers that add up to n.
}
