% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{poly_base}
\alias{poly_base}
\title{Polynomial basis vector expansion}
\usage{
poly_base(x, p)
}
\arguments{
\item{x}{a number or vector.}

\item{p}{a number (integer).}
}
\value{
polynomial basis of \code{x} up to degree \code{p}.
}
\description{
Generate polynomial basis vector up to order p.
has multivariate functionality as described in the main paper
normalized by factorials in denominator.
NOTE: currently works only up to 4th degree polynomial expansion for multivariate \code{x}.
}
\examples{
poly_base(x = 2, p = 5)
}
