% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{poly_baseC}
\alias{poly_baseC}
\title{Polynomial basis expansion in C++}
\usage{
poly_baseC(x, p)
}
\arguments{
\item{x}{a number or vector.}

\item{p}{order of polynomial expansion.}
}
\value{
vector of polynomial expansion of \code{x} up to order \code{p}.
}
\description{
polynomial basis function for univariate and multivariate input.
}
\keyword{internal}
