% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpbwcde_methods.R
\name{print.lpbwcde}
\alias{print.lpbwcde}
\title{Print Method for Local Polynomial Conditional Density Bandwidth Selection}
\usage{
\method{print}{lpbwcde}(x, ...)
}
\arguments{
\item{x}{Class "lpbwcde" object, obtained by calling \code{\link{lpbwcde}}.}

\item{...}{Other arguments.}
}
\value{
\item{Display output}{A list of specified options provided to the function.}
}
\description{
The print method for local polynomial conditional density bandwidth selection objects.
}
\seealso{
\code{\link{lpbwcde}} for data-driven bandwidth selection.

Supported methods: \code{\link{coef.lpbwcde}}, \code{\link{print.lpbwcde}}, \code{\link{summary.lpbwcde}}.
}
\author{
Matias D. Cattaneo, Princeton University. \email{cattaneo@princeton.edu}.

Rajita Chandak (maintainer), Princeton University. \email{rchandak@princeton.edu}.

Michael Jansson, University of California Berkeley. \email{mjansson@econ.berkeley.edu}.

Xinwei Ma, University of California San Diego. \email{x1ma@ucsd.edu}.
}
