% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpcde_methods.R
\name{print.lpcde}
\alias{print.lpcde}
\title{Print method for local polynomial conditional density estimation}
\usage{
\method{print}{lpcde}(x, ...)
}
\arguments{
\item{x}{Class "lpcde" object, obtained from calling \code{\link{lpcde}}.}

\item{...}{Additional options.}
}
\value{
\item{Display output}{summary of inputs to \code{lpcde}}
}
\description{
The print method for local polynomial conditional density objects.
}
\examples{
n <- 100
x_data <- as.matrix(rnorm(n, mean = 0, sd = 1))
y_data <- as.matrix(rnorm(n, mean = 0, sd = 1))
y_grid <- stats::quantile(y_data, seq(from = 0.1, to = 0.9, by = 0.1))
# density estimation
model1 <- lpcde::lpcde(x_data = x_data, y_data = y_data, y_grid = y_grid, x = 0, bw = 0.5)
print(model1)

}
\seealso{
\code{\link{lpcde}} for local polynomial conditional density estimation.
Supported methods: \code{\link{coef.lpcde}}, \code{\link{confint.lpcde}},
\code{\link{plot.lpcde}}, \code{\link{print.lpcde}},
\code{\link{summary.lpcde}}, \code{\link{vcov.lpcde}}
}
\author{
Matias D. Cattaneo, Princeton University. \email{cattaneo@princeton.edu}.

Rajita Chandak (maintainer), Princeton University. \email{rchandak@princeton.edu}.

Michael Jansson, University of California Berkeley. \email{mjansson@econ.berkeley.edu}.

Xinwei Ma, University of California San Diego. \email{x1ma@ucsd.edu}.
}
