% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{print_all_sumC}
\alias{print_all_sumC}
\title{All Sums in C++ (Internal Function)}
\usage{
print_all_sumC(target)
}
\arguments{
\item{target}{target value for sum.}
}
\value{
list of combinations that add up to target value.
}
\description{
Function that prints all combinations of natural numbers that
add up to target value.
}
\keyword{internal}
