% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpbwcde_fns.R
\name{S_exact}
\alias{S_exact}
\title{S matrix
Internal Function}
\usage{
S_exact(lower = -1, upper = 1, eval_pt, p, kernel_type)
}
\arguments{
\item{eval_pt}{evaluation point}

\item{p}{Nonnegative integer, polynomial order.}

\item{kernel_type}{String, kernel type}
}
\value{
a (p+1)-by-(p+1) matrix
}
\description{
Generate Matrix
}
\keyword{internal}
