% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{S_x}
\alias{S_x}
\title{Sx Matrix (Internal Function)}
\usage{
S_x(x_data, q, kernel_type)
}
\arguments{
\item{x_data}{Data of covariates.}

\item{q}{Maximum degree for x.}

\item{kernel_type}{Type of kernel function.}
}
\value{
S_x matrix
}
\description{
S_x matrix.
}
\keyword{internal}
