% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SxmatC}
\alias{SxmatC}
\title{S matrix in C++ (Internal Function)}
\usage{
SxmatC(x_data, q, kernel_type)
}
\arguments{
\item{x_data}{raw data vector or matrix.}

\item{q}{a number.}

\item{kernel_type}{type of kernel used}

\item{x}{evaluation point.}

\item{bw}{banwidth.}
}
\value{
Sx matrix.
}
\description{
Sx matrix pull from R.
}
\keyword{internal}
