% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpbwcde_fns.R
\name{T_y_exact}
\alias{T_y_exact}
\title{T matrix
Internal Function
Generate Matrix}
\usage{
T_y_exact(lower = -1, upper = 1, eval_pt, p, kernel_type = "uniform")
}
\arguments{
\item{eval_pt}{evaluation point}

\item{p}{Nonnegative integer, polynomial order.}

\item{kernel_type}{String, kernel type}
}
\value{
a (p+1)-by-(p+1) matrix
}
\description{
T matrix
Internal Function
Generate Matrix
}
\keyword{internal}
