% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpcde_fns.R
\name{b_x}
\alias{b_x}
\title{bx}
\usage{
b_x(datavec, s_mat, e_vec, q, kernel_type)
}
\arguments{
\item{datavec}{Dataset, vector.}

\item{s_mat}{S_hat matrix.}

\item{q}{Polynomial order.}

\item{kernel_type}{Kernel function choice.}
}
\value{
Vector of products for each data point.
}
\description{
Function for estimating the constants in the estimation formula
}
\keyword{internal}
