% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{c_x}
\alias{c_x}
\title{c_x vector (Internal Function)}
\usage{
c_x(x_data, eval_pt, m, q, h, kernel_type)
}
\arguments{
\item{x_data}{Data of covariates.}

\item{eval_pt}{Evaluation point.}

\item{m}{Order of polynomial.}

\item{q}{Maximum degree for x.}

\item{h}{Bandwidth.}

\item{kernel_type}{Type of kernel function.}
}
\description{
c_x vector generated as described in main paper.
}
\keyword{internal}
