% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_inv}
\alias{check_inv}
\title{Matrix invertibility check}
\usage{
check_inv(m)
}
\arguments{
\item{m}{Matrix}
}
\value{
TRUE if matrix is invertible.
}
\description{
Function to check intertibility of matrix.
}
\keyword{internal}
