% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpbwcde_methods.R
\name{coef.lpbwcde}
\alias{coef.lpbwcde}
\title{Coef Method for Local Polynomial Density Bandwidth Selection}
\usage{
\method{coef}{lpbwcde}(object, ...)
}
\arguments{
\item{object}{Class "lpbwcde" object, obtained by calling \code{\link{lpbwcde}}.}

\item{...}{Other arguments.}
}
\value{
\item{Estimates}{A matrix containing y_grid points and selected bandwidths.}
}
\description{
The coef method for local polynomial density bandwidth selection objects.
}
\seealso{
\code{\link{lpbwcde}} for data-driven bandwidth selection.

Supported methods: \code{\link{coef.lpbwcde}}, \code{\link{print.lpbwcde}}, \code{\link{summary.lpbwcde}}.
}
\author{
Matias D. Cattaneo, Princeton University. \email{cattaneo@princeton.edu}.

Rajita Chandak (maintainer), Princeton University. \email{rchandak@princeton.edu}.

Michael Jansson, University of California Berkeley. \email{mjansson@econ.berkeley.edu}.

Xinwei Ma, University of California San Diego. \email{x1ma@ucsd.edu}.
}
