% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpcde_fns.R
\name{cov_hat}
\alias{cov_hat}
\title{cov_hat: covariance estimator}
\usage{
cov_hat(x_data, y_data, x, y_grid, p, q, mu, nu, h, kernel_type)
}
\arguments{
\item{x_data}{covariate dataset, vector or matrix.}

\item{y_data}{response variable dataset, vector.}

\item{x}{Numeric vector or matrix, specifies the grid of evaluation points
along x-direction.}

\item{y_grid}{Numeric vector, specifies the grid of evaluation points along
y-direction.}

\item{p}{polynomial order for y.}

\item{q}{polynomial order for covariates.}

\item{mu}{degree of derivative with respect to y.}

\item{nu}{degree of derivative with respect to x.}

\item{h}{Numeric, bandwidth vector.}

\item{kernel_type}{kernel function choice.}
}
\value{
covariance matrix for all the grid points
}
\description{
Function for estimating the variance-covariance matrix.
}
\keyword{internal}
