% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kernel_eval}
\alias{kernel_eval}
\title{Kernel Evaluation function (Internal Function)}
\usage{
kernel_eval(x, kernel_type)
}
\arguments{
\item{x}{evaluation point.}

\item{kernel_type}{type of kernel function. Choose from "uniform", "triangular", "epanechnikov".}
}
\value{
kernel evaluated at \code{x}.
}
\description{
Function that evaluates product kernel at x based on the chosen function.
}
\keyword{internal}
