% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpbwcde_methods.R
\name{summary.lpbwcde}
\alias{summary.lpbwcde}
\title{Summary Method for Local Polynomial Conditional Density Bandwidth Selection}
\usage{
\method{summary}{lpbwcde}(object, ...)
}
\arguments{
\item{object}{Class "lpbwcde" object, obtained by calling \code{\link{lpbwcde}}.}

\item{...}{Additional options, including (i) \code{y_grid} specifies a subset of y_grid points
to display the bandwidth; (ii) \code{gridIndex} specifies the indices of y_grid points
to display the bandwidth.}
}
\description{
The summary method for local polynomial conditional density bandwidth selection objects.
}
\seealso{
\code{\link{lpbwcde}} for data-driven bandwidth selection.

Supported methods: \code{\link{coef.lpbwcde}}, \code{\link{print.lpbwcde}}, \code{\link{summary.lpbwcde}}.
}
\author{
Matias D. Cattaneo, Princeton University. \email{cattaneo@princeton.edu}.

Rajita Chandak (maintainer), Princeton University. \email{rchandak@princeton.edu}.

Michael Jansson, University of California Berkeley. \email{mjansson@econ.berkeley.edu}.

Xinwei Ma, University of California San Diego. \email{x1ma@ucsd.edu}.
}
