% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{T_x}
\alias{T_x}
\title{T_x matrix (Internal Function)}
\usage{
T_x(x_data, eval_pt, q, h, kernel_type)
}
\arguments{
\item{x_data}{data of covariates.}

\item{eval_pt}{evaluation point.}

\item{q}{polynomial order for covariates.}

\item{h}{bandwidth.}

\item{kernel_type}{type of kernel function.}
}
\value{
matrix.
}
\description{
constructing the Tx matrix as described in the supplemental appendix.
}
\keyword{internal}
