% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{T_x}
\alias{T_x}
\title{T_x matrix (Internal Function)}
\usage{
T_x(x_data, eval_pt, q, h, kernel_type)
}
\arguments{
\item{x_data}{Data of covariates.}

\item{eval_pt}{Evaluation point.}

\item{q}{Polynomial order for covariates.}

\item{h}{Bandwidth.}

\item{kernel_type}{Type of kernel function.}
}
\value{
Matrix.
}
\description{
Constructing the Tx matrix as described in the supplemental appendix.
}
\keyword{internal}
