% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{T_y}
\alias{T_y}
\title{T_y matrix (Internal Function)}
\usage{
T_y(y_data, yp_data, p, kernel_type)
}
\arguments{
\item{y_data}{vector of data points.}

\item{p}{polynomial order.}

\item{kernel_type}{type of kernel function. Currently only works with uniform kernel.}

\item{eval_pt}{evaluation point.}

\item{h}{bandwidth.}
}
\value{
matrix Ty.
}
\description{
constructing the Ty matrix as described in the supplemental appendix.
}
\keyword{internal}
